+++ lib/white_bread/runners/utilities.ex
@@ -1,0 +1,5 @@
+defmodule WhiteBread.Runners.Utilities do
+  def apply_scenario_starting_state(feature_state, context) do
+    apply(context, :scenario_starting_state, [feature_state])
+  end
+end