+++ test/feature/finder_test.exs
@@ -3,12 +3,12 @@ defmodule WhiteBread.Feature.FinderTest do
   alias WhiteBread.Feature.Finder
 
   test "finds only the five feature files" do
-    result = Finder.find_in_path("features/")
+    result = Finder.find_in_path("features/default/")
     assert Enum.count(result) == 7
   end
 
   test "accepts a list of paths" do
-    result = Finder.find_in_path(["features/sub_folder/", "features/sub_folder_two/"])
+    result = Finder.find_in_path(["features/default/sub_folder/", "features/default/sub_folder_two/"])
     assert Enum.count(result) == 2
   end